+++ src/lib.rs
@@ -1,7 +1,7 @@
 //! HX711 embedded-hal SPI driver crate
 //!
 //! This is a platform agnostic driver to interface with the HX711 load cell IC. It uses SPI instad of bit banging.
-//! This driver is built using [`embedded-hal`][2] traits.
+//! This driver [no_std] is built using [`embedded-hal`][2] traits.
 //!
 //!
 //! # Usage
@@ -9,17 +9,26 @@
 //! is the only device on the bus. Connect the SDO to the PD_SCK and SDI to DOUT of the HX711. SPI
 //!  clock frequency has to be between 20 kHz and 5 MHz.
 //!
+//! # Examples
 //! ```rust
-//! use rppal::spi::{Spi, Bus, SlaveSelect, Mode};
-//! use hx711_spi::{Hx711, Mode};
+//! // embedded_hal implementation
+//! use rppal::{spi::{Spi, Bus, SlaveSelect, Mode, Error},hal::Delay};
 //!
-//! let spi = Spi::new(Bus::Spi0, SlaveSelect::Ss0, 1_000_000, Mode::Mode0).unwrap();
+//! use hx711_spi::Hx711;
+//! use nb::block;
 //!
-//! // to create sensor with default configuration:
-//! let mut scale = Hx711(spi);
+//! // minimal example
+//! fn main() -> Result<(), Error>
+//! {
+//!     let spi = Spi::new(Bus::Spi0, SlaveSelect::Ss0, 1_000_000, Mode::Mode0)?;
+//!     let mut hx711 = Hx711::new(spi, Delay::new());
 //!
-//! // start measurements
-//! let mut value = scale.retrieve().unwrap();
+//! 	hx711.reset()?;
+//!     let v = block!(hx711.read())?;
+//! 	println!("value = {}", v);
+//!
+//!     Ok(())
+//! }
 //! ```
 //!
 //! # References
@@ -42,7 +51,7 @@ use hal::blocking::spi;
 use hal::blocking::delay::DelayMs;
 use core::unimplemented;
 use core::marker::Sync;
-use nb;
+use nb::{self, block};
 
 // use bitmach to decode the result
 use bitmatch::bitmatch;
@@ -64,7 +73,7 @@ pub enum Mode{
 
 /// Represents an instance of a HX711 device
 #[derive(Debug)]
-pub struct Hx711<SPI, T>
+pub struct Hx711<SPI, D>
 //where
 //    SPI: spi::Transfer<u8, Error=E> + spi::Write<u8, Error=E>,
 //    T: DelayUs<u16> + DelayMs<u16>
@@ -74,32 +83,39 @@ pub struct Hx711<SPI, T>
     // device specific
     mode: Mode,
     // timeer for delay
-    timer: T
+    delay: D
 }
 
-impl <SPI, E, T> Hx711<SPI, T>
+impl <SPI, E, D> Hx711<SPI, D>
 where
     SPI: spi::Transfer<u8, Error=E> + spi::Write<u8, Error=E>,
-    T: DelayMs<u16>
+    D: DelayMs<u16>
 {
-    /// opens a connection to a HX711 on a specified SPI
+    /// opens a connection to a HX711 on a specified SPI.
     /// The datasheet specifies PD_SCK high time and PD_SCK low time to be in the 0.2 to 50 us range,
     /// therefore bus speed has to be between 5 MHz and 20 kHz. 1 MHz seems to be a good choice.
-    // e. g. let dev = Spi::new(bus, SlaveSelect::Ss0, 1_000_000, Mode::Mode0)?;
-    pub fn new(spi: SPI, timer: T) -> Result<Self, E>
+    /// e. g. let dev = Spi::new(bus, SlaveSelect::Ss0, 1_000_000, Mode::Mode0)?;
+    /// D is an embedded_hal implementation of DelayMs
+    ///
+    /// # Safety
+    ///
+    /// It's unsafe to use Hx711 in multi-threading environments since a call to the read and reset
+    /// functions would result in undefined behaviour if the previous call has not finished first
+    /// Changing the mode is safe since it is applied on the next read and takes effect on the
+    /// second read operation.
+    pub fn new(spi: SPI, delay: D) -> Self
     {
-        Ok
-        (
-            Hx711
-            {
-                spi,
-                mode: Mode::ChAGain128,
-                timer
-            }
-        )
+        Hx711
+        {
+            spi,
+            mode: Mode::ChAGain128,
+            delay
+        }
     }
 
     /// reads a value from the HX711 and retrurns it
+    /// # Errors
+    /// Returns SPI errors and nb::Error::WouldBlock if data isn't ready to be read from hx711
     pub fn read(&mut self) -> nb::Result<i32, E>
     {
         // check if data is ready
@@ -113,7 +129,7 @@ where
         if txrx[0] == 0xFF                      // as soon as a single bit is low data is ready
         {
             // sleep for 1 millisecond which is 1/100 of the conversion period to grab the data while it's hot
-            self.timer.delay_ms(1);              // not sure if that's ok with nb
+            self.delay.delay_ms(1);              // not sure if that's ok with nb
             return Err(nb::Error::WouldBlock);
         }
 
@@ -129,19 +145,9 @@ where
         Ok(decode_output(&buffer))
     }
 
-    /// alias for read()
-    pub fn readout(&mut self) -> nb::Result<i32, E>
-    {
-        self.read()
-    }
-
-    /// alias for read()
-    pub fn retrieve(&mut self) -> nb::Result<i32, E>
-    {
-        self.read()
-    }
-
-    /// Reset the chip to it's default state. Mode is set to Convet channel A with a gain factor of 128.
+    /// Reset the chip to it's default state. Mode is set to convert channel A with a gain factor of 128.
+    /// # Errors
+    /// Returns SPI errors
     pub fn reset(&mut self) -> Result<(), E>
     {
         // when PD_SCK pin changes from low to high and stays at high for longer than 60µs,
@@ -164,13 +170,14 @@ where
     pub fn set_mode(&mut self, m: Mode) -> Result<Mode, E>
     {
         self.mode = m;
+        block!(self.read())?;           // read writes Mode for the next read()
         Ok(m)
     }
 
     /// Get the mode currently set.
-    pub fn get_mode(&mut self) -> Result<Mode, E>
+    pub fn mode(&mut self) -> Mode
     {
-        Ok(self.mode)
+        self.mode
     }
 
     /// To power down the chip the PD_SCK line has to be held in a 'high' state. To do this we
@@ -197,7 +204,7 @@ where
 }
 
 // it's not safe to use SPI bus from different treads and therefore the Hx711 driver is not
-// tread safe either
+// tread-safe either
 // this should not be necessary since the actual implementations for SPI should correctly implement Sync
 // but I don't want Sync to be auto implemented
 impl <SPI, E, T> !Sync for Hx711<SPI, T>