+++ CHANGELOG.md
@@ -13,16 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -- enable()
 -- disable()
 -- retrieve() as an alias to read()
-- [no-std]
+- [no_std]
 - nb for read()
 - CHANGELOG
 ### Changed
-- Due to the change to [no-std] the interface of new() had to change: it needs a Delay parameter now.
+- Due to the change to [no_std] the interface of new() had to change: it needs a Delay parameter now.
 - Due to nb the return type of read() / readout() has changed to nb::Result
 - readout() was changed to read()
 - readout() as alias to read()
 ## [0.2.1] - 2021-04-14
 Initial release
 
-[Unreleased]: https://github.com/crjeder/hx711_spi/blob/no-std/
+[Unreleased]: https://github.com/crjeder/hx711_spi/blob/no_std/
 [0.2.1]: https://github.com/crjeder/hx711_spi/tree/0.2.1